Remove conversion and add unequality comparison operator
The conversion operator causes CachedVar to be used as a T value,
but this is not consistent with the general use of the class.

To be consistent we should also reimplement comparison operators
that doens't check for prev value. Preferring force user to
choose what to do, i.e. specify whether compare current() or both values.

Added unequality operator between CachedVar and T template type